Update code to correct syntax
Mistakenly did not add in the && operators for comparison. Additionally,
this problem could be solved with returning a letter in a string based
on the result of 'Math.ceil/5'. For sake of the problem, and to practice
correct syntax of if/else statements, the problem was solved with the
longer version.